fixed64: codemod pass 12 -- arg-casts + ambiguous-= in #included files
The per-TU codemods only touch the compiled file, so #included actor
sources (a_iceguy, a_strifeitems, EV_DoDoor/P_LineAttack callers, ...)
kept their Fixed-arg and Fixed=angle issues. A dump-driven pass (one
tree-wide compile, then apply by the error's file regardless of which TU
surfaced it) casts Fixed function/ctor args to the param type and wraps
ambiguous Fixed=angle/double RHS in fixed_t(...). 26 sites across 12
files, behavior-preserving. Flag-off green; tests green.

+++ src/zandronum/src/g_doom/a_archvile.cpp
@@ -46,7 +46,7 @@ DEFINE_ACTION_FUNCTION_PARAMS(AActor, A_Fire)
 	ACTION_PARAM_START(1);
 	ACTION_PARAM_FIXED(height,0);
 	
-	A_Fire(self, height);
+	A_Fire(self,(int)( height));
 }
 
 void A_Fire(AActor *self, int height)
